--- Modules/HR/Database/Seeders/HrChannelSeederTableSeeder.php
@@ -1,34 +1,0 @@
-<?php
-
-namespace Modules\HR\Database\Seeders;
-
-use Carbon\Carbon;
-use Illuminate\Database\Seeder;
-use Illuminate\Support\Facades\DB;
-
-class HrChannelSeederTableSeeder extends Seeder
-{
-    /**
-     * Run the database seeds.
-     *
-     * @return void
-     */
-    public function run()
-    {
-        DB::table('hr_channel')->insert([
-            [
-                'name'=>'LinkedIn',
-                'created_at' => Carbon::now()->format('Y-m-d H:i:s'),
-                'updated_at' => Carbon::now()->format('Y-m-d H:i:s')
-            ], [
-                'name'=>'Indeed',
-                'created_at' => Carbon::now()->format('Y-m-d H:i:s'),
-                'updated_at' => Carbon::now()->format('Y-m-d H:i:s')
-            ], [
-                'name'=>'Website',
-                'created_at' => Carbon::now()->format('Y-m-d H:i:s'),
-                'updated_at' => Carbon::now()->format('Y-m-d H:i:s')
-            ]
-        ]);
-    }
-}
