+++ Modules/HR/Database/Seeders/HrChannelsTableSeeder.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace Modules\HR\Database\Seeders;
+
+use Carbon\Carbon;
+use Illuminate\Database\Seeder;
+use Illuminate\Support\Facades\DB;
+
+class HrChannelsTableSeeder extends Seeder
+{
+    /**
+     * Run the database seeds.
+     *
+     * @return void
+     */
+    public function run()
+    {
+
+        $count = DB::table('hr_channels')->count();
+        if ($count == 0) {
+            DB::table('hr_channels')->insert([
+                [
+                    'name'=>'LinkedIn',
+                    'created_at' => Carbon::now()->format('Y-m-d H:i:s'),
+                    'updated_at' => Carbon::now()->format('Y-m-d H:i:s')
+                ], [
+                    'name'=>'Indeed',
+                    'created_at' => Carbon::now()->format('Y-m-d H:i:s'),
+                    'updated_at' => Carbon::now()->format('Y-m-d H:i:s')
+                ], [
+                    'name'=>'Website',
+                    'created_at' => Carbon::now()->format('Y-m-d H:i:s'),
+                    'updated_at' => Carbon::now()->format('Y-m-d H:i:s')
+                ]
+            ]);
+        }
+        else{
+            // data already exist...
+        }
+
+    }
+}
